setup: add ase and asap3 dependencies
ASE and ASAP are the two main library frameworks used for performing MD
simulations in this project. Add them as dependencies.

Modify the installation instruction to include manually installing numpy
before installing this project. This is needed due to ASAP having
broken dependencies by needing numpy to be installed before it itself
can be installed.

The right thing to do here would be filing a problem report to the asap3
developer to have the dependency problem fixed.

Bump the required python version to 3.7 since this is what ASAP
requires.