Add running of pytest to CI
Unit tests are meant to be placed in the
salsa_dancing_molecules/unit_test directory and to be written using the
pytest library. This choice is made since pytest is the industry
standard way of testing python code.

An exclude is added to the setup script to avoid any possibility of
installing the tests.