materialsproject: add client for fetching materials data
This provides a class called MatClient that allows materials to be
queried from materialsproject.org if one has an API key. It downloads
materials containing elements given by chemical formula and converts the
result to ASE atom objects.

The minimum python version required is bumped to 3.8 since this is what
the materialsproject API library requires.