materialsproject: add function for downloading and pickling materials
As part of the design of running the program in parallel, it was decided
that downloaded materials shall be saved to disk as pickled Python
objects for later use. This adds a function to MatClient that take the
formula for a material and a path to an output directory and does
exactly this.