Create startup module for startup script generation
Create a module that will be used to generate a startup script as well
as simulation configuration for the simulation worker processes to use.
A human-readable configruation will be converted into several
json-configurations that represent a single simulation run.

Each simulation run will consume a json file and perform the simulation.