cli, materialsproject, startup: allow materialsproject materials in startup script
Add functionality to allow for materialsproject materials to be used in
startup script generation. Materials prefixed with "mp_" will be
interpreted as chemical formulas to be downloaded from Materialsproject.
Each material configuration downloaded is inserted into the simulation
configuration and has a corresponding worker process json configuration
generated for it.