optimade: rewrite to keep simulation data as attributes in structure
It turned out that the python tools do not implement the calculations
endpoint. Move all our calculated values into the structure endpoint
instead.

Remove the _salsa_ prefix from the values as the prefix is configured in
the optimade python tools. Make the default values for non-existent
fields be None, this will make filtering work correctly in the optimade
client.

The --base-url command line option is removed since there is no need for
URL references between records with all values in the same record.